better tests for Arweave and `parse_input`
